Refactor banners and health page logic.
Extract common banner component and create dedicated health page banner. Move health page logic to a new file for better organization. Adjust imports to reflect new structure.
