+++ keep-ui/components/banners/BannerBase.tsx
@@ -2,35 +2,45 @@
 import React from "react";
 import { Text, Button } from "@tremor/react";
 import Image from "next/image";
-import KeepPng from "../keep.png";
+import KeepPng from "../../keep.png";
 import { capture } from "@/shared/lib/capture";
 
-const ReadOnlyBanner = () => {
+type KeepBannerProps = {
+  bannerId: string;
+  text: string | React.ReactElement<any, any>;
+  newWindow: boolean;
+}
+
+const KeepBanner = ({
+  bannerId,
+  text,
+  newWindow = false,
+}: KeepBannerProps)  => {
   return (
     <div className="w-full py-2 pl-4 pr-2 mb-4 bg-orange-50 border border-orange-200 rounded-lg">
       <div className="flex items-center justify-between gap-4">
-        <Text className="text-sm font-medium text-black">
-          <Image
-            src={KeepPng}
-            alt="Keep Logo"
-            width={20}
-            height={20}
-            className="inline-block mr-2"
-          />
-          Keep is in read-only mode.
+        <Image
+          src={KeepPng}
+          alt="Keep Logo"
+          width={20}
+          height={20}
+          className="inline-block mr-2"
+        />
+        <Text className="text-sm font-medium text-black flex-grow">
+          {text}
         </Text>
         <div className="flex items-center gap-2">
           <Button
             className="[&>span]:text-xs"
             onClick={() => {
               capture("star-us", {
-                source: "read-only-banner",
+                source: bannerId,
               });
-              window.open(
+              {newWindow ? window.open(
                 "https://www.github.com/keephq/keep",
                 "_blank",
                 "noopener,noreferrer"
-              );
+              ) : window.location.href = "https://www.github.com/keephq/keep"}
             }}
             variant="primary"
             color="orange"
@@ -42,13 +52,13 @@ const ReadOnlyBanner = () => {
             className="[&>span]:text-xs"
             onClick={() => {
               capture("talk-to-us", {
-                source: "read-only-banner",
+                source: bannerId,
               });
-              window.open(
+              {newWindow ? window.open(
                 "https://www.keephq.dev/meet-keep",
                 "_blank",
                 "noopener,noreferrer"
-              );
+              ) : window.location.href = "https://www.keephq.dev/meet-keep"}
             }}
             color="orange"
             variant="secondary"
@@ -62,4 +72,4 @@ const ReadOnlyBanner = () => {
   );
 };
 
-export default ReadOnlyBanner;
+export default KeepBanner;
